Move ETS6 and Jetto system models to own submodules
