+++ src/duqtools/ets/__init__.py
@@ -1,7 +1,9 @@
 """This module contains tools for interfacing with ets runs."""
 
+from ._schema import Ets6SystemModel
 from ._system import Ets6System
 
 __all__ = [
     'Ets6System',
+    'Ets6SystemModel',
 ]
