+++ src/duqtools/ets/_schema.py
@@ -1,0 +1,20 @@
+from typing import Literal
+
+from pydantic import Field
+
+from ..schema import SystemModel
+
+
+class Ets6SystemModel(SystemModel):
+    name: Literal['ets6'] = Field(
+        'ets6', description='Backend system to use. Set by ConfigModel.')
+
+    submit_script_name: str = 'run.sh'
+
+    kepler_module: str = Field(
+        description='module name used in kepler load <module>')
+    kepler_load: str = Field(
+        description='module name used in kepler load <module>')
+    ets_xml: str = Field(
+        '$ITMWORK/ETS6.xml',
+        description='ETS6.XML file to use, can include for example `$ITMWORK`')
