+++ src/duqtools/jetto/__init__.py
@@ -2,6 +2,7 @@
 
 from ._batchfile import write_batchfile
 from ._jettovar_to_json import jettovar_to_json
+from ._schema import JettoSystemModel
 from ._system import BaseJettoSystem, JettoSystem, JettoSystemV210921, JettoSystemV220922
 
 __all__ = [
@@ -9,6 +10,7 @@ __all__ = [
     'JettoSystem',
     'JettoSystemV210921',
     'JettoSystemV220922',
+    'JettoSystemModel',
     'jettovar_to_json',
     'write_batchfile',
 ]
